requirements.txt
faiss_cpu==1.12.0
faiss_gpu==1.7.2
numpy==2.3.2
pandas==2.3.2
rdkit==2025.3.5
streamlit==1.48.1
torch==2.8.0
